+++ supabase/functions/invite-learner/index.ts
@@ -1,4 +1,5 @@
-import "jsr:@supabase/functions-js/edge-runtime.d.ts"
+import "jsr:@supabase/functions-js/edge-runtime.d.ts";
+import { z } from "npm:zod";
 import EdgeFunctionRequest, { RunParams } from "../_shared/EdgeFunctionRequest.ts";
 import inviteLearner from "./inviteLearner.ts";
